split and delete PureconfSuite
