+++ core/src/main/scala/pureconfig/ConfigConvert.scala
@@ -352,7 +352,7 @@ object ConfigConvert extends LowPriorityConfigConvertImplicits {
 
     override def to(t: Option[T]): ConfigValue = t match {
       case Some(v) => conv.value.to(v)
-      case None => ConfigValueFactory.fromMap(Map().asJava)
+      case None => ConfigValueFactory.fromAnyRef(null)
     }
 
     def toOption(t: Option[T]): Option[ConfigValue] = t.map(conv.value.to)
